+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Automatically import Legato snippets from leaf packages
 - Add welcome page
 - Add profile sync informatin in status bar + fix action
+- Add new Legato hierarchical view
+- Add actions to create/rename/remove apps and systems
+- Invite user to select or create a system when active definition file is not set in a Legato workspace
 
 ## 0.4.0
 - New build task to trigger build+install in one step